pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "semanticdog"
version = "0.1.0"
description = "File semantic integrity validator — fills the gap ZFS checksums miss"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "rawpy>=0.21",
    "pillow-heif>=0.16",
    "pypdf>=4.0",
    "olefile>=0.47",
    "mutagen>=1.47",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "mcp>=1.0",
    "prometheus-client>=0.20",
    "psutil>=5.9",
    "pebble>=5.0",
    "wcmatch>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[project.scripts]
sdog = "semanticdog.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning:PIL",
    "ignore::DeprecationWarning:multiprocessing",
]

[tool.hatch.build.targets.wheel]
packages = ["semanticdog"]
